move mark handle to core/handler
